add AbstractEventStoreQueryTest
